+++ examples/classifier.c
@@ -88,7 +88,7 @@ void train_classifier(char *datacfg, char *cfgfile, char *weightfile, int *gpus,
 
     int count = 0;
     int epoch = (*net->seen)/N;
-    while(get_current_batch(net) < net->max_batches || net->max_batches == 0)
+    while((int)get_current_batch(net) < net->max_batches || net->max_batches == 0)
     {
         if(net->random && count++%40 == 0)
         {
@@ -141,7 +141,7 @@ void train_classifier(char *datacfg, char *cfgfile, char *weightfile, int *gpus,
         avg_loss = avg_loss*.9 + loss*.1;
         printf("%ld, %.3f: %f, %f avg, %f rate, %lf seconds, %ld images\n", get_current_batch(net), (float)(*net->seen)/N, loss, avg_loss, get_current_rate(net), what_time_is_it_now()-time, *net->seen);
         free_data(train);
-        if(*net->seen/N > epoch)
+        if((int)*net->seen/N > epoch)
         {
             epoch = *net->seen/N;
             char buff[256];
@@ -946,7 +946,7 @@ void gun_classifier(char *datacfg, char *cfgfile, char *weightfile, int cam_inde
         printf("\033[1;1H");
 
         int threat = 0;
-        for(i = 0; i < sizeof(bad_cats)/sizeof(bad_cats[0]); ++i)
+        for(i = 0; i < (int)(sizeof(bad_cats)/sizeof(bad_cats[0])); ++i)
         {
             int index = bad_cats[i];
             if(predictions[index] > .01)
@@ -957,7 +957,7 @@ void gun_classifier(char *datacfg, char *cfgfile, char *weightfile, int cam_inde
             }
         }
         if(!threat) printf("Scanning...\n");
-        for(i = 0; i < sizeof(bad_cats)/sizeof(bad_cats[0]); ++i)
+        for(i = 0; i < (int)(sizeof(bad_cats)/sizeof(bad_cats[0])); ++i)
         {
             int index = bad_cats[i];
             if(predictions[index] > .01)
